Split DebugView into separate section files
